Move __gb_read, __gb_write to mmu
