+++ gameboy.c
@@ -2,6 +2,7 @@
 #include "gameboy.h"
 #include "apu.h"
 #include "cpu.h"
+#include "mmu.h"
 
 /* Tick the internal RTC by one second */
 void gb_tick_rtc(struct gb_s *gb)
